requirements.txt
torch
torchvision
torchaudio
transformers
datasets
scikit-learn
pandas
accelerate
fastapi
uvicorn[standard]
python-dotenv
pydantic
